+++ eurocropsml/dataset/dataset.py
@@ -65,12 +65,6 @@ class EuroCropsDataset(Dataset[LabelledData]):
         self.pad_seq_to_366 = pad_seq_to_366
         self.padding_value = padding_value
 
-        if self.config.normalize is False:
-            logger.warning(
-                "You have removed deactivate the default normalization. "
-                "This requires custom modification of sequence padding."
-            )
-
         if "S2" in self.config.satellite:
             band_names = cast(list[str], self.config.s2_bands)