+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Changes from previous releases are listed below.
 - Enhancing download process _(see #58)_
 - Load S2 split from Zenodo for benchmark _(see #59)_
 - Swap lat and lon coordinates for Eurocrops _(see #82)_
+- Move normaliztion logger warning to EuroCropsDatasetConfig _(see #84)_
 
 ## 0.3.1 (2024-07-29)
 - Remove country_code variable in collector downloader _(see #33)_